+++ tools/builder.ts
@@ -27,6 +27,8 @@ const build = async (watch: boolean) => {
       bundle: true,
       // minify: true,
       outdir: './build/src',
+      platform: 'node',
+      format: 'esm',
     });
     await buildContext.watch();
   } else {
@@ -35,6 +37,8 @@ const build = async (watch: boolean) => {
       bundle: true,
       // minify: true,
       outdir: './build/src',
+      platform: 'node',
+      format: 'esm',
     });
 
     if (buildStatus.errors.length > 0) {
